+++ src/Padoru.OfflineBans/Commands/Modify.cs
@@ -3,6 +3,7 @@
     using CommandSystem;
     using Newtonsoft.Json;
     using Padoru.OffBans.Classes;
+    using Padoru.OfflineBans.Classes;
     using System;
     using System.IO;
     using System.Linq;
@@ -10,34 +11,44 @@
 
     [CommandHandler(typeof(RemoteAdminCommandHandler))]
     [CommandHandler(typeof(GameConsoleCommandHandler))]
-    public class Modify : RecieverParent
+    public class Modify : ParentCommand
     {
-        public override string Command { get; } = "ofban modify";
+        public Modify() => LoadGeneratedCommands();
 
-        public override string[] Aliases { get; } = new string[] { "ofb modify", "offban modify" };
+        public override string Command { get; } = "modify";
 
-        public override string Description { get; } = "Редактирует офбан.";
+        public override string[] Aliases { get; } = new string[] { "edit", "mod" };
 
-        public override bool Execute(ArraySegment<string> arguments, ICommandSender sender, out string response)
+        public override string Description { get; } = "Редактирует офбан, нарушителя и нужную информацию в нужном месте.";
+
+        public override void LoadGeneratedCommands() { }
+
+        protected override bool ExecuteParent(ArraySegment<string> arguments, ICommandSender sender, out string response)
         {
+            if (!((CommandSender)sender).CheckPermission(PlayerPermissions.BanningUpToDay))
+            {
+                response = "Недостаточно прав";
+                return false;
+            }
+            
             string clearString = string.Join(" ", arguments.Array);
-            if (Regex[0].IsMatch(clearString) || Regex[1].IsMatch(clearString))
+            if (Tools.Regex[0].IsMatch(clearString) || Tools.Regex[1].IsMatch(clearString))
             {
                 response = "Формат команды:\nofban (add/modify/del) (айди нарушителя) (срок) (причина)";
                 return false;
             }
 
-            if (!Directory.Exists(filepath))
+            if (!Directory.Exists(Tools.filepath))
             {
-                Directory.CreateDirectory(filepath);
+                Directory.CreateDirectory(Tools.filepath);
             }
 
             string id = arguments.ElementAt(0);
             string reason = string.Join(" ", arguments.Skip(2).ToArray());
-            (long, string) bantime = GetBanTime(arguments.ElementAt(1));
+            (long, string) bantime = Tools.GetBanTime(arguments.ElementAt(1));
             WantedUser user = new WantedUser(id, bantime.Item1, reason);
             string json = JsonConvert.SerializeObject(user, Formatting.Indented);
-            File.WriteAllText(filepath + $"\\{id}.json", json);
+            File.WriteAllText(Tools.filepath + $"\\{id}.json", json);
 
             response = $"Бан успешно изменён:\nID нарушителя: {id}\nДлительность: {bantime.Item2}.\nПричина: {reason}";
             return false;
