+++ src/Padoru.OfflineBans/PlayerScanner.cs
@@ -5,8 +5,6 @@
     using Padoru.OfflineBans.Classes;
     using PluginAPI.Core.Attributes;
     using PluginAPI.Enums;
-    using PluginAPI.Helpers;
-    using System;
     using System.IO;
     using PlayerAPI = PluginAPI.Core.Player;
     public class PlayerScanner : IEventHandler
@@ -15,7 +13,7 @@
         public void OnPlayerJoined(PlayerAPI player)
         {
 
-            if (File.Exists(WantedPath.filepath + $"\\{player.UserId}.json"))
+            if (File.Exists(Tools.filepath + $"\\{player.UserId}.json"))
             {
                 WantedUser.Ban(player);
             }
